backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
geopandas
shapely
rasterio
scipy
python-dotenv
